+++ source/Public/Asset/Set-IMAsset.ps1
@@ -105,7 +105,6 @@
         [Parameter()]
         [string]
         $RemoveTag
-
 )
 
     BEGIN