+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Set-IMAlbum; Added parameter order
+
 ## [1.0.0] - 2024-03-20
 
 ### Added